Added a new python server to access remote prometheus server:
- moved the old server in its new own repository
- created a new server which collects datas from the federate endpoint
of a prometheus distant server

For now we don't override labels with the scrape values. Because if in
the remote cluster, both IP and PORT are the same than in our cluster
then the Host Variable of Grafana would not differentiate them.
It causes the Host Variable to have the IP and PORT of the python bridge
server.