Fixed possible ghost datas when removing a cluster / projects
If we tried to delete a project (or cluster) just before the update we
would still have its datas in the Gauge but not its occurence in the
dict => it latest datas would last forever.

The new dictionnary management is a bit heavier but does deal with that
issue.

Details about implementation :
~ n clusters
~ m projects

Rather than updating every time all the datas relative to the occurences
of the DB (costs o(n*m)) and just after rewrite all the combination of
the dict anyway, we check in o(m + n) if this is needed. As deleting of
clusters / projects happens quit rarely compared to the scrape freq, we
should gain a lot.