GitHub repository pagination
Previsously the UI for creating a new Project became unusable when the
user had access to too many GitHub repositories. To avoid that the
project list is not loaded all in once but instead page by page. To
further improve performance, the pagination api endpoint only sends
fields actually needed for display instead of the full GitHub api
result.